server/requirements.txt
beanie
fastapi[standard]
pymongo
python-dotenv
loguru
uvicorn
pandas
openpyxl
python-multipart
httpx
aiohttp
litellm

gradio